Migrate language structures to a dedicated crate.
