+++ rs-parser/Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 
 members = [
+	"language",
 	"parser",
 	"postprocessor",
 	"checker",
