+++ rs-parser/app/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+language = { path = "../language" }
 parser = { path = "../parser" }
 postprocessor = { path = "../postprocessor" }
 checker = { path = "../checker" }
