+++ rs-parser/language/Cargo.toml
@@ -1,9 +1,7 @@
 [package]
-name = "parser"
+name = "language"
 version = "0.2.0"
 authors = ["Kineolyan <kineolyan@protonmail.com>"]
 edition = "2018"
 
 [dependencies]
-language = { path = "../language" }
-nom = "6.0.0"
