+++ rs-parser/parser/src/address.rs
@@ -1,73 +1,8 @@
-use std::fmt;
-
 use nom::IResult;
+use language::address::{Node, Port};
 
-use nom::number::complete::be_u32;
+// use nom::number::complete::be_u32;
 // use common::to_string;
-
-#[derive(PartialEq)]
-pub enum Node {
-	In,
-	Out,
-	Node(String),
-}
-
-impl fmt::Debug for Node {
-	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-		self.do_fmt(f)
-	}
-}
-
-impl fmt::Display for Node {
-	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-		self.do_fmt(f)
-	}
-}
-
-impl Node {
-	#[cfg(test)]
-	pub fn new_node(name: &str) -> Self {
-		Node::Node(name.to_string())
-	}
-
-	pub fn get_id<'a>(&'a self) -> &'a String {
-		match self {
-			&Node::Node(ref id) => id,
-			_ => panic!("Not a named node: {}", self),
-		}
-	}
-
-	fn do_fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-		match self {
-			&Node::In => write!(f, "<IN>"),
-			&Node::Out => write!(f, "<OUT>"),
-			&Node::Node(ref id) => write!(f, "Node#{}", id),
-		}
-	}
-}
-
-#[derive(Debug, PartialEq)]
-pub struct Port {
-	pub node: Node,
-	pub port: u32,
-}
-
-impl Port {
-	pub fn new(node: Node, port: u32) -> Self {
-		Port {
-			node: node,
-			port: port,
-		}
-	}
-
-	#[cfg(test)]
-	pub fn named_port(node_name: &str, port: u32) -> Self {
-		Port {
-			node: Node::new_node(node_name),
-			port: port,
-		}
-	}
-}
 
 fn input_node(input: &[u8]) -> IResult<&[u8], Node> {
 	let (remaining, _) = nom::bytes::complete::tag("IN")(input)?;
